+++ libs/ui/u-ui/src/lib/components/address-form/address-form.component.ts
@@ -1,4 +1,4 @@
-import { ChangeDetectorRef, Component, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, ChangeDetectorRef, Component, OnInit } from '@angular/core';
 import { Address, Select } from '@ivt/c-data';
 import { filter, startWith, switchMap, takeUntil } from 'rxjs/operators';
 
@@ -9,8 +9,7 @@ import { createAddressForm, IvtAddressFormService } from './address-form.service
   selector: 'ivt-address-form',
   templateUrl: './address-form.component.html',
   styleUrls: ['./address-form.component.scss'],
-  // Don't use change detection for displaying error messages
-  // changeDetection: ChangeDetectionStrategy.OnPush,
+  changeDetection: ChangeDetectionStrategy.OnPush,
   providers: [IvtAddressFormService],
 })
 export class IvtAddressFormComponent extends IvtFormComponent<Address> implements OnInit {
@@ -50,5 +49,9 @@ export class IvtAddressFormComponent extends IvtFormComponent<Address> implement
         this.suburbOptions = suburbOptions;
         this.cdr.detectChanges();
       });
+  }
+
+  markForCheck(): void {
+    this.cdr.markForCheck();
   }
 }