+++ libs/ui/u-ui/src/lib/components/form/form.component.ts
@@ -1,7 +1,8 @@
 import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
 import { FormGroup } from '@angular/forms';
+import { Subject } from 'rxjs';
 
-import { collectFormErrors, markFormGroupTouched } from '../../functions';
+import { collectFormErrors } from '../../functions';
 import { IvtSubscriberComponent } from '../subscriber';
 
 @Component({
@@ -14,6 +15,7 @@ export class IvtFormComponent<T> extends IvtSubscriberComponent {
   @Input() item: T;
   @Input() loading: boolean;
   @Input() isEditable = true;
+  protected stateChanged = new Subject<void>();
   @Output() submitForm = new EventEmitter<T>();
 
   get values(): T {
@@ -24,12 +26,13 @@ export class IvtFormComponent<T> extends IvtSubscriberComponent {
     if (this.form.valid || this.form.disabled) {
       this.submitForm.emit(this.values);
     } else {
-      markFormGroupTouched(this.form);
+      this.form.markAllAsTouched();
       console.info('Tried to submit form with errors:', {
         form: this.form.errors,
 
         controls: collectFormErrors(this.form),
       });
     }
+    this.stateChanged.next();
   }
 }